+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+## [0.3.0] - 2025-03-26
 
-- Better code, I have no idea tbh
+### Changed
+
+- Change message handling from raw Bukkit API to CommandAPI
+- Use modern Java features to make code more readable
+- Use MiniMessage for message formatting
 
 ## [0.2.0] - 2025-01-27
 
@@ -25,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 [unreleased]: https://github.com/walker84837/simplewarp/compare/v0.2.0...HEAD
+[0.3.0]: https://github.com/walker84837/simplewarp/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/walker84837/simplewarp/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/walker84837/simplewarp/releases/tag/v0.1.0